Discrete Solution reworked

+++ ugbase/lib_discretization/reference_element/reference_elements.h
@@ -110,7 +110,7 @@ class BaseReferenceElement {
 
 class ReferenceTriangle{
 	public:
-		static const size_t dim = 2;
+		static const int dim = 2;
 
 
 		ReferenceTriangle()
